requirements.txt
pdfplumber>=0.10.0
python-docx>=1.0.0
pytesseract>=0.3.10
openpyxl>=3.1.0
Pillow>=10.0.0
ttkthemes>=3.2.2
